Yh/post listand comment

+++ src/main/java/est/commitdate/controller/BoardController.java
@@ -2,6 +2,9 @@ package est.commitdate.controller;
 
 import est.commitdate.dto.board.BoardDto;
 import est.commitdate.service.BoardService;
+import est.commitdate.service.SwipeService;
+import est.commitdate.service.member.MemberService;
+import jakarta.servlet.http.HttpSession;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Controller;
@@ -14,6 +17,8 @@ import org.springframework.web.bind.annotation.*;
 @RequestMapping("/board")
 public class BoardController {
     private final BoardService boardService;
+    private final MemberService memberService;
+
     public String prefix = "view/board/";
 
     @GetMapping("")
@@ -24,9 +29,14 @@ public class BoardController {
 
     // 게시판 생성 화면
     @GetMapping("/saveView")
-    public String boardSaveView(Model model) {
-        model.addAttribute("boardDto", new BoardDto());
-        return prefix +"boardSave";
+    public String boardSaveView(Model model, HttpSession session) {
+        //관리자라면
+        if(memberService.AuthorizationCheck(session)) {
+            model.addAttribute("boardDto", new BoardDto());
+            return prefix +"boardSave";
+        }
+        //관리자가 아니라면
+        return "redirect:/board";
     }
     // 생성 요청을 처리
     @PostMapping("/save")
@@ -37,29 +47,43 @@ public class BoardController {
 
     // 게시판 업데이트 화면
     @GetMapping("/updateView/{id}")
-    public String boardUpdateView(@PathVariable Integer id, Model model) {
-        model.addAttribute("boardDto", BoardDto.from(boardService.getBoardById(id)));
-        return "view/board/boardUpdate";
+    public String boardUpdateView(@PathVariable Integer id, Model model, HttpSession session) {
+        //관리자라면
+        if(memberService.AuthorizationCheck(session)) {
+            model.addAttribute("boardDto", BoardDto.from(boardService.getBoardById(id)));
+            return "view/board/boardUpdate";
+        }
+        //관리자가 아니라면
+        return "redirect:/board";
     }
 
     // 게시판 업데이트
     @PostMapping("/update")
     public String boardUpdateReq(BoardDto boardDto) {
-        log.info("update요청호출"+ boardDto.toString());
         boardService.update(boardDto);
         return "redirect:/board";
     }
     // 게시판 삭제
     @PostMapping("/delete/{id}")
-    public String deleteBoard(@PathVariable Integer id) {
-        boardService.delete(id);
+    public String deleteBoard(@PathVariable Integer id , HttpSession session) {
+        //관리자라면
+        if(memberService.AuthorizationCheck(session)) {
+            boardService.delete(id);
+            return "redirect:/board";
+        }
+        //관리자가 아니라면
         return "redirect:/board";
     }
 
     // 게시판 복구
     @PostMapping("/restore/{id}")
-    public String boardRestore(@PathVariable Integer id) {
-        boardService.restore(id);
-        return "redirect:/board/saveView";
+    public String boardRestore(@PathVariable Integer id ,HttpSession session) {
+        //관리자라면
+        if(memberService.AuthorizationCheck(session)) {
+            boardService.restore(id);
+            return "redirect:/board";
+        }
+        //관리자가 아니라면
+        return "redirect:/board";
     }
 }
